move the interacive chart around as well
