+++ web/src/routes/experiments/[experimentId]/interactive-chart.svelte
